Split file and common models
