+++ django-rgd/rgd/admin/__init__.py
@@ -1,4 +1,5 @@
 from . import mixins  # noqa
 from .collection import *  # noqa
 from .common import *  # noqa
+from .file import *  # noqa
 from .fileset import *  # noqa
