+++ django-rgd/rgd/models/__init__.py
@@ -1,12 +1,7 @@
 from .collection import Collection, CollectionPermission  # noqa
-from .common import (  # noqa
-    ChecksumFile,
-    FileSourceType,
-    SpatialAsset,
-    SpatialEntry,
-    WhitelistedEmail,
-)
+from .common import SpatialAsset, SpatialEntry, WhitelistedEmail  # noqa
 from .constants import *  # noqa
+from .file import ChecksumFile, FileSourceType  # noqa
 from .fileset import FileSet  # noqa
 from .mixins import *  # noqa
 from .transform import transform_geometry  # noqa
