+++ django-rgd/rgd/models/file.py
@@ -1,5 +1,4 @@
 from contextlib import contextmanager
-import json
 import logging
 import os
 from pathlib import Path
@@ -11,7 +10,6 @@ from django.conf import settings
 from django.contrib.gis.db import models
 from django_extensions.db.models import TimeStampedModel
 from filelock import FileLock
-from model_utils.managers import InheritanceManager
 from rgd.utility import (
     _link_url,
     clean_file_cache,
@@ -31,58 +29,10 @@ from s3_file_field import S3FileField
 
 from .. import tasks
 from .collection import Collection
-from .constants import DB_SRID
 from .fileset import FileSet
 from .mixins import PermissionPathMixin, TaskEventMixin
 
 logger = logging.getLogger(__name__)
-
-
-class SpatialEntry(models.Model):
-    """Common model to all geospatial data entries.
-
-    This is intended to be used in a mixin manner.
-
-    """
-
-    # `InheritanceManager` allows us to select inhereted tables via `objects.select_subclasses()`
-    objects = InheritanceManager()
-
-    spatial_id = models.AutoField(primary_key=True)
-
-    # Datetime of creation for the dataset
-    acquisition_date = models.DateTimeField(null=True, default=None, blank=True)
-
-    # This can be used with GeoDjango's geographic database functions for spatial indexing
-    footprint = models.GeometryField(srid=DB_SRID)
-    outline = models.GeometryField(srid=DB_SRID)
-
-    instrumentation = models.CharField(
-        max_length=100,
-        null=True,
-        blank=True,
-        help_text='The instrumentation used to acquire these data.',
-    )
-
-    def __str__(self):
-        try:
-            return 'Spatial ID: {} (ID: {}, type: {})'.format(self.spatial_id, self.id, type(self))
-        except AttributeError:
-            return super().__str__()
-
-    @property
-    def bounds(self):
-        extent = {
-            'xmin': self.outline.extent[0],
-            'ymin': self.outline.extent[1],
-            'xmax': self.outline.extent[2],
-            'ymax': self.outline.extent[3],
-        }
-        return extent
-
-    @property
-    def bounds_json(self):
-        return json.dumps(self.bounds)
 
 
 class FileSourceType(models.IntegerChoices):
@@ -351,24 +301,3 @@ class ChecksumFile(TimeStampedModel, TaskEventMixin, PermissionPathMixin):
         return _link_url(self, 'get_url')
 
     data_link.allow_tags = True
-
-
-class WhitelistedEmail(models.Model):
-    """Pre-approve users for sign up by their email."""
-
-    email = models.EmailField()
-
-
-class SpatialAsset(SpatialEntry, TimeStampedModel, PermissionPathMixin):
-    """Any spatially referenced file set.
-
-    This can be any collection of files that have a spatial reference and are
-    not explicitly handled by the other SpatialEntry subtypes. For example, this
-    model can be used to hold a collection of PDF documents or slide decks that
-    have a georeference.
-
-    """
-
-    permissions_paths = [('files', ChecksumFile)]
-
-    files = models.ManyToManyField(ChecksumFile, related_name='+')
